refactor(test): concise component test setup
The goal of this refactor is to make the intention of tests
clearer and making future tests easier to implement.

Most tests using the mpc component testing framework
will focus on one specific protocol. It is useful to have
that configuration summarized as

```rust
only_generate_triples()
```

instead of copy-pasting the setup to disable key sharing and presignature generation:

```rust
.with_preshared_key()
.with_min_presignatures_stockpile(0)
.with_max_presignatures_stockpile(0)
```

The second refactor in this commit moves the waiting for RPC actions to
a reusable function.